milestone01/infra/requirements.txt
aws-cdk-lib==2.190.0
constructs==10.4.2
cdk-ecr-deployment==4.0.1
